Fixed the progress calculator as it was only showing 0%

+++ src/fitnesstrackerapp/ProgressTracker.java
@@ -20,7 +20,7 @@ public class ProgressTracker {
             return "You haven't started working out yet!";
         }
 
-        int progress = (burned / calorieGoal) * 100;
+        int progress = (int) (((double) burned / calorieGoal) * 100);
 
         if (progress >= 100) {
             return "Congratulations! You have reached your goal!";
